Starting on new crash system
Began working on a new crash system, the user will have an extra life in case they crash. I plan on making the car go into a ‘ghost’ mode and be invincible and be semi-transparent, this is because sometimes the traffic will arrange themselves in ways that are impossible to drive past.