Ghost system and time added
Fully implemented a ghost system into the game as well as a working stopwatch that tracks how long you are in the game. If you collide with a car for the first time, you will get a 3 second window to get out of danger without being able to crash, but after that you have no more chances.