Added points and increased speed every 10 seconds
Added a new point system in which 10 points are awarded every time you successfully pass a car and it goes off-screen. For next time, add a point deduction for crashing, also a high score system that could be saved as well.

Began adding a feature in which the cars will gradually become faster every 10 seconds, but needs more tweaking to make the increase noticeable.

Noticed a bug in which after a few seconds (unsure how long) of playing, the ghost mode will run permanently even after 3 seconds, and needs to be fixed.