Background downscroll and two new traffic vehicles
I extended the background and added a background scroll so it would appear as if the road was infinitely moving. I'm planning on adjusting the speed that it moves as the player moves.

Made two new vehicle sprites and added them into the game. Planning new sedans and different colors for each vehicle. I want at least 5 traffic vehicles in the game, and make it randomly change once they go off-screen.