Patched ghost bug, milestone tweaks, and pausing
Fixed the ghost bug. Found out that I accidentally made the time the ability ran the same as the time elapsed since the program began and would only disable it once the time was equal to 3. Thus if you activated the ability after 3 seconds had passed it would run infinitely. I fixed it by creating a variable inside the conditional statement handling collisions by adding 3 seconds to the time elapsed and waiting until it matched up, then deactivating it.

Adjusted the milestone to make the cars much faster every 10 seconds that pass. Also will now earn 50 points for reaching a milestone.

The feature will now have to press the spacebar in order to start the game. Planning on adding a loading screen that allows you to unlock other vehicles.

Working on adding a pause feature by pressing escape.