All cars centered, and background adjustments
Made sure every car would be centered in their respective lane, adjusting for the different dimensions of each car.

Made the traffic faster but the background slower. I honestly am now thinking all of the things moving at once becomes an eye sore to look at and I’m not quite sure how I should move along with it. So I’m working on a mechanic where the background moves slowly but will rapidly increase moving in the opposite y-direction as the player.