high score system and began working on restarting
Polished the high score system. I had it read a file and check if it was empty to see if a high score had been set or not. If it was empty then any score would be accepted as a high score, but from then on the high score would have to be higher than the previous.

Began working on a way to restart the game after crashing, currently testing to see which variables should be re-initialized in order to start the game efficiently and functionally.