Ghost code completed
Got a bulk of the coding for the ghost powerup completed. It will spawn within the first 30 seconds then 20-30 seconds after that. But I plan on rotating different power ups to have variety.


Once the ghosts spawn time syncs with the time elapsed, the ghost will scroll down the screen on either of the 3 lanes until either the user picks it up or it falls out of the screen. Then it will spawn 20-30 seconds later.

Began working on merging two of the on screen menu items together under one class