Hourglass power-up
Merged the stats board and controls under one class, planning on doing the same for future on screen graphics

Added an hourglass power-up that will slow down traffic. Right now it activates as soon as you activate the power-up but I plan on adding a feature where you can enable it whenever you want.

Re-initializing many of the ghosts variables. From now on I want to have as many variables under classes as possible. As for the variables for the ghost I’m somewhat done, but there are still a few yet to be completed