Polishing up classes and main.py
Finished moving all of the variables for the ghost power up to its class. Did the same for the users car, moving the milestone and points variable into the class, this is all so the code can be more organized and main.py is less crowded. I can also just add a function in each class to easily keep track of and initialize all the variables when I want to restart the game.

Began testing all of my traffic as different parts of one sprite group, rather than 3 separate classes. So far it works perfectly fine as stand-ins for the traffic, but for some reason when I use the .add() function it flags an error in pygame but it has seemingly no influence on the actual game and works normally